Move mapper functions to their own file
